% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startMlrPlayground.R
\name{start}
\alias{start}
\title{Start mlrPlayground app}
\usage{
start(...)
}
\arguments{
\item{...}{[\code{any}]\cr
Additional arguments passed to shiny's
\code{runApp()} function.}
}
\description{
Run a local instance of mlrPlayground
}
\examples{
\dontrun{
  start()
}
}
